q_rand_walk.m
Plots the PDF and frequency distribution of a quantum random walk upto 'n' steps, without measurement at immediate steps, using a Hadamard coin. Results of measurment at the last step, i.e. the collapse to one of the '2*n+1' steps to the left or right, are probabilistic and can be inferred from the plotted PDF. To emulate the measurement process, we use random number generation in tandem with the CDF of the generated PDF, to plot a frequency distribution. Refer to the paper Quantum Random Walks by J. Kempe (Fig 5 for the PDF).